Once again completely restructured most things.
I went with a bootstrap template due to my art skills in front end web design lacking. also setup the backend to hook the openIMDB api via a model that I exported to my main startup node.js script through a module and logged the result. Next step after this is to send the data back to the client and parse the data in controller.

Matthew Richardson.