Pre getting google Oauth2.0 to handle user data.
Setting up Oauth 2.0 to handle user data reqires the site to be hosted and that data is nessisary to be saved in the database as a user id for storage. this update is the routes to google oauth and the callback as well as a few minor changes to selected movie and data passthrough.